Build KWalletManager ourselves
The Flatpak isn't suitable for our use cases since it doesn't (and
can't) include the KCM that's also in the source repo, and we want
that to be accessible in System Settings.

Let's make sure to build it from source here, and then we can stop
pre-installing the Flatpak version.