Don't build dolphin
The intention here was to have Dolphin from Flatpak, but it gets built
by accident as a part of the `dolphin-plugins` build plan. However we
only need those plugins for the context menu on the Plasma desktop, and
not Dolphin itself.

Blacklist Dolphin from building so it doesn't get pulled into the base
image by accident.