Use our own custom kde-builder.yaml file as the source of truth
Instead of generating a standard one and then overriding a bunch of
stuff in it, let's use a custom one with the options we want.

In addition to greater correctness, this lets us offer the config file
publicly to developers wanting to use kde-builder on KDE Linux.

PKGBUILD generation is unchanged; now relevant data is read from the
kde-builder.yaml file instead of make-pkgbuilds.py

Resolves https://invent.kde.org/kde-linux/kde-linux-packages/-/issues/2
Related to https://invent.kde.org/kde-linux/kde-linux/-/issues/213