Revert "Compile Calamares here"
This reverts commit c959f4d13bddd5a78ffe961f71c454018211d6c7.

This won't work. It won't create a PKGBUILD for it because, Calamares
being a 3rd-party package, nothing is wired up for this. Calamares is
already in AUR, so the low-friction thing to do would be to get it from
there, but we don't want to introduce new AUR dependencies, so that's
not a good option.

Reverting this until we can do a larger-scale re-think of how the
packages pipeline works.